Allow Parity as ingest node  (#36)
* Upgrade go-ethereum to v1.8

* Add Node Info for parity nodes

* Upgrade start_private_blockchain to use v1.8

+++ pkg/geth/block_rewards.go
@@ -28,7 +28,7 @@ func calcTransactionFees(block core.Block) float64 {
 	var transactionFees float64
 	for _, transaction := range block.Transactions {
 		receipt := transaction.Receipt
-		transactionFees += float64(transaction.GasPrice * receipt.GasUsed)
+		transactionFees += float64(uint64(transaction.GasPrice) * receipt.GasUsed)
 	}
 	return transactionFees / params.Ether
 }
